+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - portscan: print progress reports on `SIGINFO` or `SIGUSR2` or in
   regular intervals when requested with `-o progress`
+- portscan: Report commented `PORTEPOCH` or `PORTREVISION` lines
+  via new `lint.commented-portrevision` lint, selectable with
+  `-o comments`, enabled by default
 
 ### Changed